+++ src/main/java/syncsquad/teamsync/logic/parser/AddMeetingCommandParser.java
@@ -4,7 +4,7 @@ import static syncsquad.teamsync.logic.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 
 import syncsquad.teamsync.logic.commands.AddMeetingCommand;
 import syncsquad.teamsync.logic.parser.exceptions.ParseException;
-import syncsquad.teamsync.model.schedule.Meeting;
+import syncsquad.teamsync.model.meeting.Meeting;
 
 /**
  * Parses input arguments and creates a new AddMeetingCommand object
