+++ src/main/java/syncsquad/teamsync/model/ReadOnlyAddressBook.java
@@ -2,7 +2,7 @@ package syncsquad.teamsync.model;
 
 import javafx.collections.ObservableList;
 import syncsquad.teamsync.model.person.Person;
-import syncsquad.teamsync.model.schedule.Meeting;
+import syncsquad.teamsync.model.meeting.Meeting;
 
 /**
  * Unmodifiable view of an address book
