+++ src/main/java/syncsquad/teamsync/model/meeting/exceptions/DuplicateMeetingException.java
@@ -1,4 +1,4 @@
-package syncsquad.teamsync.model.schedule.exceptions;
+package syncsquad.teamsync.model.meeting.exceptions;
 
 /**
  * Signals that the operation will result in duplicate Meetings (Meetings are considered duplicates if they have the
